Week 3, Lesson 1, Exercise 1: Angular ui-router and Resolve
Objectives and Outcomes

In this exercise you will use resolve in ui-router to simplify the code within your controllers. At the end of this exercise, you will be able to:

- Use the resolve object to resolve the dependencies in the state
- Inject the data after resolution into the controller

Conclusions

In this exercise, you updated the app to use the resolve in ui-router to simplify the controllers.

Changes to be committed:
	modified:   www/js/app.js
	modified:   www/js/controllers.js
	modified:   www/js/services.js