Week 3, Lesson 1, Exercise 2: Displaying Loading Message During State Transition
Objectives and Outcomes

In this exercise you will learn to make use of the support for events in Angular to trigger the showing and hiding of a loading message in your application. At the end of this exercise, you will be able to:

- Use $rootScope to create a mechanism to display and hide loading message
- Use the ui-router events to show and hide the loading message

Conclusions

In this exercise, you have learnt about watching for events and triggering some action in response to the events.

Changes to be committed:
	modified:   www/js/app.js