Week 3, Lesson 2, Exercise 1 - Using Local Storage
Objectives and Outcomes

In this exercise, you will learn about HTML5 local storage and how you can leverage this to support the storing of information in your app in local storage. At the end of this exercise, you will be able to:

- Design a new Angular service using local storage
- Make use of the Local Storage to remember user information

Conclusions

In this exercise, you learnt to make use of the HTML5 local storage within your application, first by creating a service to wrap the local storage, and then inject into a controller in order to use the storage.

Changes to be committed:
	modified:   www/js/controllers.js
	modified:   www/js/services.js