Week 3, Lesson 5, Assignment 1, Task 1: Peer Graded Assignment: Resolve and Local Storage
Instructions
In this assignment you will be concentrating on using ui-router's resolve object to update the app. In addition you will use the local storage to persist the information about the user's favorite dishes. The three tasks to be accomplished in this assignment are as follows:

Task 1

- Update the home, menu and aboutus states in app.js to use resolve to support data resolution in the ui-router.

Changes to be committed:
	modified:   www/js/app.js
	modified:   www/js/controllers.js