Week 3, Lesson 5, Assignment 3: Peer Graded Assignment: Resolve and Local Storage Instructions
In this assignment you will be concentrating on using ui-router's resolve object to update the app. In addition you will use the local storage to persist the information about the user's favorite dishes. The three tasks to be accomplished in this assignment are as follows:

Task 1
- Update the home, menu and aboutus states in app.js to use resolve to support data resolution in the ui-router.

Task 2
- Update IndexController, MenuController and AboutController to inject the data obtained through resolution and simplify the controller code

Task 3
- Make use of local storage support in the app to persist the user's favorite dishes on the device. The information should persist across app shut down and restart.
Changes to be committed:
	modified:   www/js/controllers.js
	modified:   www/js/services.js