Week 4, Lesson 1, Exercise 1: Customizing and Controlling the Splash Screen
Objectives and Outcomes

In this exercise you will learn how to customize the splash screen and also control the duration of its display on the screen. To do this, you will make use of Cordova through the ngCordova splash screen support. At the end of this exercise, you will be able to:

- Customize the splash screen of an Ionic application
- Control the display of the splash screen through the Cordova plugin support

  Conclusions

  In this exercise, you learnt the use of your first ngCordova and Cordova plugin in order to customize and control the splash screen.