Week 4, Lesson 2, Exercise 1: Notifying the User
Objectives and Outcomes

In this exercise, you will use two different Cordova plugins for delivering notification to the user. You will use the local notification plugin to put a notification into the device's notification bar. In addition you will use the toast plugin to show a short message on the screen to alert the user. At the end of this exercise, you will be able to:

- Use the Cordova Local Notification plugin together with the ngCordova wrapper to put notifications into the device's notification bar
- Use the Cordova Toast plugin together with the ngCordova wrapper to show a short message to the user on the device's screen

Conclusions

In this exercise, you learnt to use the Cordova local notifications and
toast plugins to notify the user.

Changes to be committed:
	modified:   package.json
	modified:   www/js/controllers.js