Week 4, Lesson 3, Exercise 1: Using the Camera Plugin
Objectives and Outcomes

In this exercise, you will continue exploring more Cordova plugins. In particular, you will use the Cordova camera plugin to access the device's built-in camera to retrieve image data and use it within your application.  At the end of this exercise, you will be able to:

- Use the Cordova camera plugin together with the ngCordova wrapper to access the device's native camera to retrieve image data
- Make use of the image data within your application

Changes to be committed:
	new file:   www/templates/register.html